# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: AnyLive API
  version: 0.1.0
  description: |
    AnyLive overseas-first live streaming control plane.
    P0 surface: system health + meta. Auth/rooms freeze in v0.1/v0.2.
servers:
  - url: http://localhost:8088
    description: Local API
...
paths:
...
  /api/v1/rooms:
    post:
      tags: [rooms]
      operationId: createRoom
      summary: Create a room (P1)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoomRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "401":
          $ref: "#/components/responses/Error"
    get:
      tags: [rooms]
      operationId: listRooms
      summary: List live/hot rooms (P1)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [idle, live, closed]
      responses:
        "200":
          description: Room list
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Room"
  /api/v1/rooms/{roomId}:
    get:
      tags: [rooms]
      operationId: getRoom
      summary: Get room by id
      parameters:
        - $ref: "#/components/parameters/RoomId"
      responses:
        "200":
          description: Room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "404":
          $ref: "#/components/responses/Error"
...
  /api/v1/rooms/{roomId}/start:
    post:
      tags: [rooms]
      operationId: startRoom
      security: [{bearerAuth: []}]
      parameters:
        - $ref: "#/components/parameters/RoomId"
      responses:
        "200": {description: Live}
  /api/v1/rooms/{roomId}/stop:
    post:
      tags: [rooms]
      operationId: stopRoom
      security: [{bearerAuth: []}]
      parameters:
        - $ref: "#/components/parameters/RoomId"
      responses:
        "200": {description: Idle}
...
  /api/v1/rooms/{roomId}/media/publish:
    post:
      tags: [rooms]
      operationId: mediaPublish
      security: [{bearerAuth: []}]
      parameters:
        - $ref: "#/components/parameters/RoomId"
      responses:
        "200": {description: Publish info}
  /api/v1/rooms/{roomId}/media/play:
    get:
      tags: [rooms]
      operationId: mediaPlay
      parameters:
        - $ref: "#/components/parameters/RoomId"
      responses:
        "200": {description: Play URLs}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RoomId:
      name: roomId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
...
    ApiError:
      type: object
      required: [code, message, retryable]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
        retryable:
          type: boolean
...
    CreateRoomRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 80
...
    Room:
      type: object
      required: [id, owner_id, title, status, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        title:
          type: string
        status:
          type: string
          enum: [idle, live, closed]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...